% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_destination}
\alias{plan_destination}
\title{Add a destination to a data plan}
\usage{
plan_destination(plan_obj, dst)
}
\arguments{
\item{plan_obj}{a \code{data_plan} object.}

\item{dst}{a URI}
}
\value{
\code{plan_obj}
}
\description{
Add a destination to a data plan
}
