% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_reader}
\alias{plan_reader}
\title{Add a reader to a data plan}
\usage{
plan_reader(plan_obj, reader, ...)
}
\arguments{
\item{plan_obj}{a \code{data_plan} object.}

\item{reader}{a function that takes a character object as its first parameter}

\item{...}{additional parameters passed to \code{reader} when executed}
}
\value{
\code{plan_obj}
}
\description{
Add a reader to a data plan
}
