% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_source}
\alias{plan_source}
\title{Add a source to a data plan}
\usage{
plan_source(plan_obj, src)
}
\arguments{
\item{plan_obj}{a \code{data_plan} object.}

\item{src}{a URI}
}
\value{
\code{plan_obj}
}
\description{
Add a source to a data plan
}
