% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_transform}
\alias{plan_transform}
\title{Add a transformation to a data plan}
\usage{
plan_transform(plan_obj, fn, ...)
}
\arguments{
\item{plan_obj}{a \code{data_plan} object.}

\item{fn}{a function that takes data as the first parameter}

\item{...}{additional parameters passed to \code{fn} when executed}
}
\value{
\code{plan_obj}
}
\description{
Add a transformation to a data plan
}
