% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_writer}
\alias{plan_writer}
\title{Add a writer to a data plan}
\usage{
plan_writer(plan_obj, writer, ...)
}
\arguments{
\item{plan_obj}{a \code{data_plan} object.}

\item{writer}{a function that takes data as the first parameter
and a destination as the second}

\item{...}{additional parameters passed to \code{writer} when executed}
}
\value{
\code{plan_obj}
}
\description{
Add a writer to a data plan
}
